[MOVE/#3] 기본 UiState 패키지 이동
