+++ app/src/main/java/com/napzak/market/presentation/dummy/state/DummyUiState.kt
@@ -1,6 +1,6 @@
 package com.napzak.market.presentation.dummy.state
 
-import com.napzak.market.core.common.util.UiState
+import com.napzak.market.core.common.state.UiState
 import com.napzak.market.domain.dummy.model.DummyUser
 
 data class DummyUiState(
